+++ workspaces/features-demo/src/demo-pieces/deprecated-tests.ts
@@ -107,20 +107,20 @@ export function createDeprecatedTestsDemo() {
 
         .addMeasure()
         .addLyrics(1, "1. Any text", "1n", { align: "left" })
-        .addNote(0, "G4", "2n").addAnnotation("any text", { group: "dynamics" })
+        .addNote(0, "G4", "2n").addAnnotation("any text", "dynamics")
         .addNote(0, "G4", "2n")
         .addMeasure()
         .addLyrics(1, "2. Any text", "1n", { align: "left" })
-        .addNote(0, "G4", "2n").addAnnotation("any text", { group: "dynamics" })
+        .addNote(0, "G4", "2n").addAnnotation("any text", "dynamics")
         .addNote(0, "G4", "2n")
         .endRow()
 
         .addMeasure()
         .addLyrics(1, "1. No depr. just testing", "1n", { align: "left" })
-        .addNote(0, "G4", "4n").addAnnotation("pitchLabel", { group: "pitchLabel" })
-        .addNote(0, "G4", "4n").addAnnotation("pitchLabel", { group: "pitchLabel" })
-        .addNote(0, "G4", "4n").addAnnotation("chordLabel", { group: "chordLabel" })
-        .addNote(0, "G4", "4n").addAnnotation("chordLabel", { group: "chordLabel" })
+        .addNote(0, "G4", "4n").addAnnotation("pitchLabel", "pitchLabel")
+        .addNote(0, "G4", "4n").addAnnotation("pitchLabel", "pitchLabel")
+        .addNote(0, "G4", "4n").addAnnotation("chordLabel", "chordLabel")
+        .addNote(0, "G4", "4n").addAnnotation("chordLabel", "chordLabel")
 
         .getDocument();
 }